chore: create controller package
